Moved libraries to src folder
