--- lib/emotion-cache.ts
@@ -1,5 +1,0 @@
-import createCache from "@emotion/cache";
-
-export default createCache({
-  key: "css",
-});
