+++ src/lib/emotion-cache.ts
@@ -1,0 +1,5 @@
+import createCache from "@emotion/cache";
+
+export default createCache({
+  key: "css",
+});
