+++ src/lib/selectOptions.ts
@@ -1,0 +1,10 @@
+export const tags = [
+  { value: "react", label: "react" },
+  { value: "graphql", label: "graphql" },
+  { value: "server", label: "server" },
+  { value: "programming", label: "programming" },
+  { value: "technology", label: "technology" },
+  { value: "javascript", label: "javascript" },
+  { value: "startup", label: "startup" },
+  { value: "design", label: "design" },
+];
